+++ extensions/zendesk/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Zendesk Changelog
 
+## [Add Action to Open Article in Browser] - {PR_MERGE_DATE}
+
+- The `Open in Browser` action goes to the JSON article. There is now a new Action with shortcut `Cmd + N` that goes to the article instead.
+- Some refactoring of the code to use the `useFetch` hook
+
 ## [Updates] - 2022-12-22
 - ⬆️ Updated: Readme
 - ⬆️ Updated: all dependencies